Make Head class handle when to set gameOver to true
Before, the render method was handling when to set gameOver to true.
It should not be the render method's responsibility for handling
this logic, and it is inefficient having to keep setting gameOver to
true for every frame while the player is dead.

It is better to let the Head class handle when to set gameOver to
true because it is the one that handles the events when the player
collides with an object.